requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
colorama==0.4.6
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
qrcode==8.2